Move `sort-import-groups` to its own folder.
